feat(solver): separate knitro status enum from solver
